requirements.txt
streamlit==1.29.0
easyocr==1.7.0
opencv-python<=4.6.0.66
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
python-docx==1.1.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
requests==2.31.0
Pillow==10.1.0
plotly==5.17.0
streamlit-option-menu==0.3.6
google-generativeai==0.8.3